Разделила программу на модули
