--- lab2 var5/downloaded_images/bird_red/annotations.csv
@@ -1,2 +1,0 @@
-absolute_path,relative_path
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_red\annotations.csv,bird_red\annotations.csv
