--- lab2 var5/downloaded_images/bird_yellow/annotations.csv
@@ -1,13 +1,0 @@
-absolute_path,relative_path
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000001.jpg,bird_yellow\000001.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000002.jpg,bird_yellow\000002.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000003.jpg,bird_yellow\000003.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000004.jpg,bird_yellow\000004.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000005.png,bird_yellow\000005.png
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000006.jpg,bird_yellow\000006.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000007.jpg,bird_yellow\000007.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000008.jpg,bird_yellow\000008.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000009.jpg,bird_yellow\000009.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000010.jpg,bird_yellow\000010.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\000011.jpg,bird_yellow\000011.jpg
-C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\downloaded_images\bird_yellow\annotations.csv,bird_yellow\annotations.csv
