+++ lab2 var5/images/bird_blue/annotations.csv
@@ -1,0 +1,23 @@
+absolute_path,relative_path
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000001.jpg,bird_blue\000001.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000002.jpg,bird_blue\000002.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000003.jpg,bird_blue\000003.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000004.jpg,bird_blue\000004.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000005.jpg,bird_blue\000005.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000006.jpg,bird_blue\000006.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000007.jpg,bird_blue\000007.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000008.jpg,bird_blue\000008.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000009.jpg,bird_blue\000009.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000010.jpg,bird_blue\000010.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000011.jpg,bird_blue\000011.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000012.jpg,bird_blue\000012.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000013.jpg,bird_blue\000013.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000014.jpg,bird_blue\000014.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000015.jpg,bird_blue\000015.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000016.jpg,bird_blue\000016.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000017.jpg,bird_blue\000017.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000018.jpg,bird_blue\000018.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000019.jpg,bird_blue\000019.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000020.jpg,bird_blue\000020.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\000021.jpg,bird_blue\000021.jpg
+C:\Users\gayvo\OneDrive\Desktop\Прикладное програмирование\lab2 var5\images\bird_blue\annotations.csv,bird_blue\annotations.csv
